requirements.txt
pyqt5==5.15.4
xlrd==1.2.0
numpy==1.19.5
opencv-python==3.4.1.15
future==1.0.0
pillow==8.4.0
pymssql==2.2.7
pyodbc==4.0.39
numba==0.53.1
matplotlib==3.3.4
scipy==1.5.4
findpeaks==2.6.1
pandas==1.1.5
pyyaml==6.0.1
psutil==5.9.8
tqdm==4.64.1